ShapeFunction/CorrelationFunction reorganization
