+++ src/frontend/matchers/ObservablesMatcher.cpp
@@ -27,7 +27,7 @@
 #include "core/observables/correlation/RadialEnumerator.h"
 #include "core/observables/correlation/LayerwiseRadialEnumerator.h"
 #include "core/observables/correlation/PairAveragedCorrelation.h"
-#include "core/observables/correlation/S110Correlation.h"
+#include "core/observables/correlation_functions/S110Correlation.h"
 #include "core/observables/DensityHistogram.h"
